move ArraySpec to its own file
