+++ src/zarr/codecs/_v2.py
@@ -6,8 +6,9 @@ import numcodecs
 from numcodecs.compat import ensure_bytes, ensure_ndarray
 
 from zarr.abc.codec import ArrayArrayCodec, ArrayBytesCodec
+from zarr.array_spec import ArraySpec
 from zarr.buffer import Buffer, NDBuffer
-from zarr.common import JSON, ArraySpec, to_thread
+from zarr.common import JSON, to_thread
 
 
 @dataclass(frozen=True)
