+++ src/zarr/codecs/bytes.py
@@ -15,7 +15,8 @@ from zarr.common import parse_enum, parse_named_configuration
 if TYPE_CHECKING:
     from typing_extensions import Self
 
-    from zarr.common import JSON, ArraySpec
+    from zarr.array_spec import ArraySpec
+    from zarr.common import JSON
 
 
 class Endian(Enum):
