+++ src/zarr/codecs/crc32c_.py
@@ -14,7 +14,8 @@ from zarr.common import parse_named_configuration
 if TYPE_CHECKING:
     from typing_extensions import Self
 
-    from zarr.common import JSON, ArraySpec
+    from zarr.array_spec import ArraySpec
+    from zarr.common import JSON
 
 
 @dataclass(frozen=True)
