+++ src/zarr/codecs/pipeline.py
@@ -27,7 +27,8 @@ from zarr.metadata import ArrayMetadata
 if TYPE_CHECKING:
     from typing_extensions import Self
 
-    from zarr.common import ArraySpec, SliceSelection
+    from zarr.array_spec import ArraySpec
+    from zarr.common import SliceSelection
 
 T = TypeVar("T")
 U = TypeVar("U")
