+++ tests/v3/package_with_entrypoint/__init__.py
@@ -1,7 +1,8 @@
 from numpy import ndarray
 
 from zarr.abc.codec import ArrayBytesCodec
-from zarr.common import ArraySpec, BytesLike
+from zarr.array_spec import ArraySpec
+from zarr.common import BytesLike
 
 
 class TestCodec(ArrayBytesCodec):
